added feature modify song data

+++ data_pipeline/artist_modifications/remove_old_artist_data.sh
@@ -1,9 +1,7 @@
 #!/bin/bash
 
-#!/bin/bash
-
 # filepath from view of working directory of python interpreter
 artists_folder="data/artists"
 
 # remove folder contents
-rm -rf "$artists_folder"/*
+rm -rf "${artists_folder:?}"/*
